requirements/base.txt
django==1.11.4
mysqlclient==1.3.10
requests==2.18.3
Markdown==2.6.8
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
validate_email==1.3